+++ Runtime/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.Networking.Transport.EditorTests")]
-[assembly: InternalsVisibleTo("Unity.Networking.Transport.PlayTests")]
+[assembly: InternalsVisibleTo("Unity.Networking.Transport.RuntimeTests")]
 [assembly: InternalsVisibleTo("Unity.Networking.Transport.PlayTests.Performance")]
 
 // We are making certain things visible for certain projects that require